# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RR Plastic Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stock-in:
    get:
      operationId: listStockIn
      tags: [stock-in]
      summary: List stock in records
      security:
        - bearerAuth: []
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: materialId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of stock in records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockIn"
    post:
      operationId: createStockIn
      tags: [stock-in]
      summary: Create a stock in record
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStockInBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockIn"
  /stock-in/{id}:
    put:
      operationId: updateStockIn
      tags: [stock-in]
      summary: Update a stock in record
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStockInBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockIn"
    delete:
      operationId: deleteStockIn
      tags: [stock-in]
      summary: Delete a stock in record
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /stock-out:
    get:
      operationId: listStockOut
      tags: [stock-out]
      summary: List stock out records
      security:
        - bearerAuth: []
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: materialId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of stock out records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockOut"
    post:
      operationId: createStockOut
      tags: [stock-out]
      summary: Create a stock out record
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStockOutBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockOut"
        "400":
          description: Insufficient stock
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stock-out/{id}:
    put:
      operationId: updateStockOut
      tags: [stock-out]
      summary: Update a stock out record
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStockOutBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockOut"
    delete:
      operationId: deleteStockOut
      tags: [stock-out]
      summary: Delete a stock out record
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /production:
    get:
      operationId: listProduction
      tags: [production]
      summary: List production records
      security:
        - bearerAuth: []
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of production records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Production"
    post:
      operationId: createProduction
      tags: [production]
      summary: Create a production record
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductionBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Production"
  /production/{id}:
    put:
      operationId: updateProduction
      tags: [production]
      summary: Update a production record
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductionBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Production"
    delete:
      operationId: deleteProduction
      tags: [production]
      summary: Delete a production record
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    StockIn:
      type: object
      properties:
        id:
          type: integer
        materialId:
          type: integer
        materialName:
          type: string
        quantity:
          type: number
        date:
          type: string
        supplier:
          type: string
        remarks:
          type: string
        createdAt:
          type: string
      required: [id, materialId, materialName, quantity, date, createdAt]
    CreateStockInBody:
      type: object
      properties:
        materialId:
          type: integer
        quantity:
          type: number
        date:
          type: string
        supplier:
          type: string
        remarks:
          type: string
      required: [materialId, quantity, date]
    StockOut:
      type: object
      properties:
        id:
          type: integer
        materialId:
          type: integer
        materialName:
          type: string
        quantity:
          type: number
        date:
          type: string
        purpose:
          type: string
        remarks:
          type: string
        createdAt:
          type: string
      required: [id, materialId, materialName, quantity, date, createdAt]
    CreateStockOutBody:
      type: object
      properties:
        materialId:
          type: integer
        quantity:
          type: number
        date:
          type: string
        purpose:
          type: string
        remarks:
          type: string
      required: [materialId, quantity, date]
...
    Production:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        materialId:
          type: integer
        materialName:
          type: string
        quantityIssued:
          type: number
        quantityReceived:
          type: number
        date:
          type: string
        remarks:
          type: string
        createdAt:
          type: string
      required: [id, customerId, customerName, materialId, materialName, quantityIssued, quantityReceived, date, createdAt]
    CreateProductionBody:
      type: object
      properties:
        customerId:
          type: integer
        materialId:
          type: integer
        quantityIssued:
          type: number
        quantityReceived:
          type: number
        date:
          type: string
        remarks:
          type: string
      required: [customerId, materialId, quantityIssued, quantityReceived, date]
...